+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Rules/SubLedger.hs
@@ -200,6 +200,7 @@ instance
   , InjectRuleFailure "SUBDELEG" DijkstraSubDelegPredFailure era
   , InjectRuleFailure "SUBLEDGER" Conway.ConwayLedgerPredFailure era
   , InjectRuleFailure "SUBUTXOW" Alonzo.AlonzoUtxowPredFailure era
+  , InjectRuleFailure "SUBDELEG" Shelley.AccountAlreadyRegistered era
   , TxCert era ~ DijkstraTxCert era
   ) =>
   STS (SUBLEDGER era)
@@ -216,8 +217,8 @@ instance
 dijkstraSubLedgersTransition ::
   forall era.
   ( EraTx era
-  , ConwayEraTxBody era
   , ConwayEraGov era
+  , ConwayEraTxBody era
   , EraRule "SUBLEDGER" era ~ SUBLEDGER era
   , EraRule "SUBGOV" era ~ SUBGOV era
   , EraRule "SUBUTXOW" era ~ SUBUTXOW era