+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Delpl.hs
@@ -32,6 +32,7 @@ import Cardano.Ledger.Credential (Ptr)
 import Cardano.Ledger.Shelley.Core
 import Cardano.Ledger.Shelley.Era (DELPL, ShelleyEra)
 import Cardano.Ledger.Shelley.Rules.Deleg (
+  AccountAlreadyRegistered,
   DELEG,
   DelegEnv (..),
   ShelleyDelegEvent,
@@ -221,6 +222,8 @@ instance
   , AtMostEra "Babbage" era
   , PredicateFailure (EraRule "DELEG" era) ~ ShelleyDelegPredFailure era
   , Event (EraRule "DELEG" era) ~ ShelleyDelegEvent era
+  , EraRuleFailure "DELEG" era ~ ShelleyDelegPredFailure era
+  , InjectRuleFailure "DELEG" AccountAlreadyRegistered era
   ) =>
   Embed (DELEG era) (DELPL era)
   where