+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/ImpTest.hs
@@ -231,6 +231,7 @@ import Cardano.Ledger.Shelley.LedgerState (
   utxosUtxo,
  )
 import Cardano.Ledger.Shelley.Rules (
+  AccountAlreadyRegistered,
   BbodyEnv (..),
   LedgerEnv (..),
   ShelleyBbodyState (..),
@@ -498,6 +499,7 @@ class
   , Signal (EraRule "NEWEPOCH" era) ~ EpochNo
   , EraRuleFailure "BBODY" era ~ PredicateFailure (EraRule "BBODY" era)
   , EraRuleFailure "LEDGER" era ~ PredicateFailure (EraRule "LEDGER" era)
+  , InjectRuleFailure "LEDGER" AccountAlreadyRegistered era
   , InjectRuleFailure "LEDGER" ShelleyDelegPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyUtxowPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyUtxoPredFailure era