+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Rules/Deleg.hs
@@ -6,6 +6,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
 
 module Test.Cardano.Ledger.Shelley.Rules.Deleg (
   tests,
@@ -14,7 +15,8 @@ module Test.Cardano.Ledger.Shelley.Rules.Deleg (
 import Cardano.Ledger.Coin
 import Cardano.Ledger.Shelley (hardforkAlonzoAllowMIRTransfer)
 import Cardano.Ledger.Shelley.Core
-import Cardano.Ledger.Shelley.Rules (DELEG, DelegEnv (..))
+import Cardano.Ledger.Shelley.Rules
+import qualified Cardano.Ledger.Shelley.Rules as Shelley
 import Cardano.Ledger.Shelley.State
 import Data.Foldable (fold)
 import Data.Foldable as F (foldl')
@@ -61,6 +63,8 @@ tests ::
   , ShelleyEraAccounts era
   , ChainProperty era
   , QC.HasTrace (CHAIN era) (GenEnv MockCrypto era)
+  , EraRuleFailure "DELEG" era ~ Shelley.ShelleyDelegPredFailure era
+  , InjectRuleFailure "DELEG" Shelley.AccountAlreadyRegistered era
   ) =>
   TestTree
 tests =